refactor: decompose base input
