+++ src/components/Banner.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import Form from './Form'
+import ShortenForm from './ShortenForm'
 
 export default function Banner() {
   return (
     <div className='banner'>
       <h1>Short your link</h1>
       <h3>Make a simple, clean, short links, that are easy to post, send and share</h3>
-      <Form maxLength={120}/>
+      <ShortenForm/>
     </div>
   )
 }
